Create tests

+++ start.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-rm -rf models
-source ~/ml/ml_venv/bin/activate
+if [ "$1" == "--reset" ]; then
+    rm -rf models
+fi
+if ["$1" == "--test"]; then
+    pytest tests/ --cov=src --cov=app --cov-report=term-missing
+fi
+source SleepVenv/bin/activate
 streamlit run main.py --server.headless true
